==> nb/7.prediction.ipynb <==
import pandas as pd 
import numpy as np 
import awswrangler as wr
from datetime import date
import datetime

import mlflow
from mlflow.tracking import MlflowClient

import lightgbm as lgb
import joblib
from sklearn.metrics import roc_auc_score, confusion_matrix

import pickle 
import boto3

today = date.today()
day, month, year = (today - datetime.timedelta(days=60)).strftime("%d/%m/%Y").split('/')
period_before = year + month

day, month, year = today.strftime("%d/%m/%Y").split('/')
period = year + month

id_exp = mlflow.get_experiment_by_name('exp_6').experiment_id

df_runs = mlflow.search_runs(experiment_ids=[id_exp], filter_string="metrics.val_auc > 0.70", order_by=["metrics.val_auc DESC"], max_results=1)

s3 = boto3.resource('s3')

remote_path = df_runs.artifact_uri.values[0]
model = pickle.loads(s3.Bucket(remote_path.split('/')[2]).Object(f'{"/".join(remote_path.split("/")[3:])}/model/model.pkl').get()['Body'].read())

test = wr.s3.read_parquet("s3://rimac-analytics-temporal/individuals/Dante/cross-sell-veh/data/prec/potencial/data.parquet")


preds = model.predict(test.drop(['periodo_target', 'cuc', 'placa', 'target', 'jerarquia'], axis=1), num_iteration=model.best_iteration)
test['preds'] = preds

test.preds.median()

test[test.periodo_target == 202102].shape

test[test.periodo_target == 202102].shape

wr.s3.to_parquet(
        df=test[['periodo_target', 'cuc', 'placa', 'target', 'preds']],
        path=f's3://rimac-analytics-temporal/individuals/Dante/cross_prop_veh/predictions/proba_{period}/',
        mode='overwrite',
        dataset=True,
        database="coe_analytics_tmp",
        table=f"crossveh_test_{period}"
)

f"coe_analytics_tmp.crossveh_test_{period}"

all_data = wr.s3.read_parquet("s3://rimac-analytics-temporal/individuals/Dante/cross-sell-veh/data/prec/all/data.parquet")


preds = model.predict(all_data.drop(['periodo_target', 'cuc', 'placa', 'target', 'jerarquia'], axis=1), num_iteration=model.best_iteration)
all_data['preds'] = preds

all_data[all_data.cuc == '1-44956524']

wr.s3.to_parquet(
        df=all_data[['periodo_target', 'cuc', 'placa', 'target', 'preds']],
        path=f's3://rimac-analytics-temporal/individuals/Dante/cross_prop_veh/predictions/all/',
        mode='overwrite',
        dataset=True,
        database="coe_analytics_tmp",
        table=f"crossveh_test_back"
)

f"crossveh_test_{period}"
